Refactor LeftSidebar
